+++ NEWS.md
@@ -9,6 +9,7 @@
 * Imported settings are automatically selected in the preset input
 * Use `bs4Dash::toast` instead of native shiny notifications
 * When comparison is deactivated, reset the comparison select input
+* New notification when settings are exported to json file
   
 # TissueTMDD 0.1.0